Change build from sphinx to jupyter-book
Jupyter-book is just a better-packaged distribution of sphinx
for the dependencies they use. Wanted to upgrade since it
seems like their other extensions are built for this.

This commit is huge and likely to break, but thought one commit to
revert was better in the long run to avoid confusion.

Changes:
* Makefiles and README to use new build instructions
* Configuration files to match the jupyter-book spec
* Table of contents to the jupyter-book format
* Change GitHub actions runner to match