Fix typo, and see below for additional suggestion
The phrase "returned the case" doesn't make sense in "So if a for loop can loop over a sequence of indices returned the case, would it also be possible to just loop over the str directly?"...not sure what it's meant to be. Perhaps "in this case" rather than "the case"?